Fixed bug where a swapped frame caused an infinite loop. Changed default value in pad lookup table. Turned off missing channel warnings. Other misc fixes for int sizes.

+++ get-manip/DataFile.cpp
@@ -100,7 +100,7 @@ void DataFile::CloseFile()
     if (filestream.is_open() and this->isInitialized) filestream.close();
 }
 
-const long unsigned int DataFile::GetPosition()
+const long long unsigned int DataFile::GetPosition()
 {
     if (!isInitialized) throw Exceptions::Not_Init();
     return filestream.tellg();
